Support publishing release candidates
Note that this would be much faster (and simpler) if the sha256 file of the zip were available within the google bucket.
This generates a package whose version should be `x.y.z-rc<number>`

+++ scripts/packages/chocolatey/build.ps1
@@ -1,27 +1,47 @@
 param(
   [string] $version = "0.3.2",
-  [switch] $isRelease
+  [int] $rc = 0,
+  [string] $mode = "local",
+  [string] $checksum = ""
 )
 
-$tvVersion = $version
-$tvFilename = "bazel-$($version)-windows-x86_64.zip"
-if ($isRelease) {
-  $tvUri = "https://github.com/bazelbuild/bazel/releases/download/$($version)/$($tvFilename)"
-} else {
+if ($mode -eq "release") {
+  $tvVersion = $version
+  $tvFilename = "bazel-$($tvVersion)-windows-x86_64.zip"
+  $tvUri = "https://github.com/bazelbuild/bazel/releases/download/$($tvVersion)/$($tvFilename)"
+} elseif ($mode -eq "rc") {
+  $tvVersion = "$($version)-rc$($rc)"
+  $tvFilename = "bazel-$($version)rc$($rc)-windows-x86_64.zip"
+  $tvUri = "https://storage.googleapis.com/bazel/$($version)/rc$($rc)/$($tvFilename)"
+} elseif ($mode -eq "local") {
+  $tvVersion = $version
+  $tvFilename = "bazel-$($tvVersion)-windows-x86_64.zip"
   $tvUri = "http://localhost:8000/$($tvFilename)"
+} else {
+  write-error "mode parameter '$mode' unsupported. Please use local, rc, or release."
 }
-write-host "download uri: $($tvUri)"
-
 rm -force -ErrorAction SilentlyContinue ./*.nupkg
-rm -force -ErrorAction SilentlyContinue ./*.zip
 rm -force -ErrorAction SilentlyContinue ./bazel.nuspec
 rm -force -ErrorAction SilentlyContinue ./tools/LICENSE.txt
+rm -force -ErrorAction SilentlyContinue ./tools/params.json
+if ($checksum -ne "") {
+  rm -force -ErrorAction SilentlyContinue ./*.zip
+}
 
-if ($isRelease) {
+if ($mode -eq "release") {
   Invoke-WebRequest "$($tvUri).sha256" -UseBasicParsing -passthru -outfile sha256.txt
   $tvChecksum = (gc sha256.txt).split(' ')[0]
   rm sha256.txt
-} else {
+} elseif ($mode -eq "rc") {
+  if (-not(test-path $tvFilename)) {
+    Invoke-WebRequest "$($tvUri)" -UseBasicParsing -passthru -outfile $tvFilename
+  }
+  if ($checksum -eq "") {
+    $tvChecksum = (get-filehash $tvFilename -algorithm sha256).Hash
+  } else {
+    $tvChecksum = $checksum
+  }
+} elseif ($mode -eq "local") {
   Add-Type -A System.IO.Compression.FileSystem
   $outputDir = "$pwd/../../../output"
   $zipFile = "$pwd/$($tvFilename)"
@@ -41,5 +61,12 @@ From: https://github.com/bazelbuild/bazel/blob/master/LICENSE.txt
 "@
 add-content -value $licenseHeader -path "./tools/LICENSE.txt"
 add-content -value (get-content "../../../LICENSE.txt") -path "./tools/LICENSE.txt"
+
+$params = @{
+  uri = $tvUri;
+  checksum = $tvChecksum;
+  checksumType = "sha256"
+}
+add-content -value (ConvertTo-Json $params) -path "./tools/params.json"
 
 choco pack ./bazel.nuspec
